move basis data to output folder
