+++ source/databricks/calculation_engine/package/calculation/output/__init__.py
@@ -12,5 +12,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .energy_results import write_energy_results
-from .wholesale_results import write_wholesale_results
