+++ source/databricks/calculation_engine/package/calculation/output/basis_data/schemas/charge_link_periods_schema.py
