+++ source/databricks/calculation_engine/package/calculation/output/basis_data/schemas/charge_price_information_periods_schema.py
