+++ source/databricks/calculation_engine/package/calculation/output/basis_data/schemas/grid_loss_metering_points_schema.py
