+++ source/databricks/calculation_engine/package/calculation/output/basis_data/schemas/metering_point_period_schema.py
