+++ source/databricks/calculation_engine/package/calculation/output/basis_data/schemas/time_series_point_schema.py
