+++ source/databricks/calculation_engine/package/calculation/output/calculation_results/calculations_storage_model_factory.py
@@ -15,7 +15,7 @@ from dependency_injector.wiring import Provide
 from pyspark.sql import SparkSession, DataFrame, Row
 
 from package.calculation import PreparedDataReader
-from package.calculation.basis_data.schemas import calculations_schema
+from package.calculation.output.basis_data.schemas import calculations_schema
 from package.calculation.calculator_args import CalculatorArgs
 from package.constants.calculation_column_names import CalculationColumnNames
 from package.container import Container
