+++ source/databricks/calculation_engine/package/calculation/output/calculation_results/schemas/energy_results_schema.py
