+++ source/databricks/calculation_engine/package/calculation/output/calculation_results/schemas/total_monthly_amounts_schema.py
