+++ source/databricks/calculation_engine/package/calculation/wholesale/wholesale_calculation.py
@@ -11,11 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Tuple
 
-import package.calculation.output.wholesale_storage_model_factory as wholesale_results_factory
-import package.calculation.output.monthly_amounts_per_charge_storage_model_factory as monthly_amounts_factory
-import package.calculation.output.total_monthly_amounts_storage_model_factory as total_monthly_amounts_factory
+import package.calculation.output.calculation_results.wholesale_storage_model_factory as wholesale_results_factory
+import package.calculation.output.calculation_results.monthly_amounts_per_charge_storage_model_factory as monthly_amounts_factory
+import package.calculation.output.calculation_results.total_monthly_amounts_storage_model_factory as total_monthly_amounts_factory
 
 import package.calculation.wholesale.fee_calculators as fee_calculator
 import package.calculation.wholesale.subscription_calculators as subscription_calculator
