+++ source/databricks/calculation_engine/package/datamigration_hive/schema_config.py
@@ -13,25 +13,25 @@
 # limitations under the License.
 from spark_sql_migrations import Schema, Table, View
 
-import package.calculation.basis_data.schemas as basis_data_schemas
+import package.calculation.output.basis_data.schemas as basis_data_schemas
 import package.infrastructure.paths as paths
 
 # calculation_input
 from package.calculation.input.schemas.grid_loss_metering_points_schema import (
     grid_loss_metering_points_schema,
 )
-from package.calculation.output.schemas.energy_results_schema import (
+from package.calculation.output.calculation_results.schemas import (
     hive_energy_results_schema,
 )
-from package.calculation.output.schemas.total_monthly_amounts_schema import (
+from package.calculation.output.calculation_results.schemas import (
     total_monthly_amounts_schema,
 )
-from package.calculation.output.schemas.monthly_amounts_schema import (
+from package.calculation.output.calculation_results.schemas import (
     monthly_amounts_schema,
 )
 
 # calculation_output
-from package.calculation.output.schemas.wholesale_results_schema import (
+from package.calculation.output.calculation_results.schemas.wholesale_results_schema import (
     wholesale_results_schema,
 )
