+++ source/databricks/calculation_engine/tests/calculation/output/test_monthly_amounts_per_charge_storage_model_factory.py
@@ -20,9 +20,7 @@ import pytest
 from pyspark.sql import SparkSession, DataFrame
 
 from package.calculation.calculator_args import CalculatorArgs
-from package.calculation.output import (
-    monthly_amounts_per_charge_storage_model_factory as sut,
-)
+from package.calculation.output.calculation_results import monthly_amounts_per_charge_storage_model_factory as sut
 from package.calculation.wholesale.data_structures import MonthlyAmountPerCharge
 from package.calculation.wholesale.data_structures.monthly_amount_per_charge import (
     monthly_amount_per_charge_schema,
