+++ source/databricks/calculation_engine/tests/calculation/preparation/transformations/test_get_metering_point_periods_basis_data.py
@@ -15,10 +15,10 @@
 from pyspark.sql import SparkSession, Row
 
 from calculation.preparation.transformations import metering_point_periods_factory
-from package.calculation.basis_data.basis_data import (
+from package.calculation.output.basis_data import (
     get_metering_point_periods_basis_data,
 )
-from package.calculation.basis_data.schemas import hive_metering_point_period_schema
+from package.calculation.output.basis_data import hive_metering_point_period_schema
 from package.codelists import (
     MeteringPointResolution,
 )
