+++ source/databricks/calculation_engine/tests/calculation/preparation/transformations/test_get_time_series_basis_data.py
@@ -16,8 +16,8 @@ from pyspark.sql import SparkSession
 from calculation.preparation.transformations import (
     prepared_metering_point_time_series_factory,
 )
-from package.calculation.basis_data.schemas import time_series_point_schema
-from package.calculation.basis_data.basis_data import (
+from package.calculation.output.basis_data import time_series_point_schema
+from package.calculation.output.basis_data import (
     get_time_series_points_basis_data,
 )
 from package.common import assert_schema
